// src/types/eventCodes.ts
export enum EventCode {
  // 200 OK
  LOGIN_SUCCESS = "LOGIN_SUCCESS",
  ACCOUNT_FETCHED = "ACCOUNT_FETCHED",
  ACCOUNT_UPDATED = "ACCOUNT_UPDATED",
  ACCOUNT_CLOSED = "ACCOUNT_CLOSED",
  TRANSACTION_FETCHED = "TRANSACTION_FETCHED",
  TRANSFER_FETCHED = "TRANSFER_FETCHED",

  // 201 CREATED
  USER_REGISTERED = "USER_REGISTERED",
  ACCOUNT_CREATED = "ACCOUNT_CREATED",
  TRANSACTION_CREATED = "TRANSACTION_CREATED",
  TRANSFER_CREATED = "TRANSFER_CREATED",

  // 400 BAD REQUEST
  EMPTY_BODY = "EMPTY_BODY",
  INVALID_CUSTOMER_ID = "INVALID_CUSTOMER_ID",
  INVALID_ACCOUNT_ID = "INVALID_ACCOUNT_ID",
  MISSING_REQUIRED_FIELDS = "MISSING_REQUIRED_FIELDS",
  DUPLICATE_ACCOUNT = "DUPLICATE_ACCOUNT",
  DUPLICATE_RESOURCE = "DUPLICATE_RESOURCE",
  VALIDATION_ERROR = "VALIDATION_ERROR",

  // 401 UNAUTHORIZED
  INVALID_TOKEN = "INVALID_TOKEN",
  INVALID_CREDENTIALS = "INVALID_CREDENTIALS",

  // 403 FORBIDDEN
  FORBIDDEN = "FORBIDDEN",

  // 404 NOT FOUND
  ACCOUNT_NOT_FOUND = "ACCOUNT_NOT_FOUND",
  TRANSACTION_NOT_FOUND = "TRANSACTION_NOT_FOUND",
  TRANSFER_NOT_FOUND = "TRANSFER_NOT_FOUND",

  // 409 CONFLICT
  USERNAME_ALREADY_EXISTS = "USERNAME_ALREADY_EXISTS",
  EMAIL_ALREADY_EXISTS = "EMAIL_ALREADY_EXISTS",
  UNKNOWN_CONFLICT = "UNKNOWN_CONFLICT",
  ACCOUNT_NOT_ACTIVE = "ACCOUNT_NOT_ACTIVE",
  NO_SELF_TRANSFER_ALLOWED = "NO_SELF_TRANSFER_ALLOWED",
  INSUFFICIENT_FUNDS = "INSUFFICIENT_FUNDS",

  // 500 INTERNAL SERVER ERROR
  INTERNAL_SERVER_ERROR = "INTERNAL_SERVER_ERROR",
}